Refactor middleware functions in one middleware file
