+++ routes/comments.js
@@ -2,9 +2,10 @@ var express = require('express');
 var router = express.Router({ mergeParams: true });
 var Place = require('../models/place');
 var Comment = require('../models/comment');
+var middleware = require('../middleware'); // automaticly will look for index.js in the directory, so no need to specify
 
 // comments new route
-router.get('/new', isLoggedIn, function(req, res) {
+router.get('/new', middleware.isLoggedIn, function(req, res) {
   Place.findById(req.params.id, function(err, place) {
     if (err) {
       console.log(err);
@@ -15,7 +16,7 @@ router.get('/new', isLoggedIn, function(req, res) {
 });
 
 // comments create route
-router.post('/', isLoggedIn, function(req, res) {
+router.post('/', middleware.isLoggedIn, function(req, res) {
   // lookup place using id
   Place.findById(req.params.id, function(err, place) {
     if (err) {
@@ -44,7 +45,10 @@ router.post('/', isLoggedIn, function(req, res) {
 });
 
 // comments edit route
-router.get('/:comment_id/edit', checkCommentOwnership, function(req, res) {
+router.get('/:comment_id/edit', middleware.checkCommentOwnership, function(
+  req,
+  res
+) {
   Comment.findById(req.params.comment_id, function(err, foundComment) {
     if (err) {
       res.redirect('back');
@@ -58,7 +62,10 @@ router.get('/:comment_id/edit', checkCommentOwnership, function(req, res) {
 });
 
 // comment update route
-router.put('/:comment_id', checkCommentOwnership, function(req, res) {
+router.put('/:comment_id', middleware.checkCommentOwnership, function(
+  req,
+  res
+) {
   Comment.findByIdAndUpdate(req.params.comment_id, req.body.comment, function(
     err,
     updatedComment
@@ -72,7 +79,10 @@ router.put('/:comment_id', checkCommentOwnership, function(req, res) {
 });
 
 // comment delete route
-router.delete('/:comment_id', checkCommentOwnership, function(req, res) {
+router.delete('/:comment_id', middleware.checkCommentOwnership, function(
+  req,
+  res
+) {
   // find by id and remove
   Comment.findByIdAndRemove(req.params.comment_id, function(err) {
     if (err) {
@@ -82,34 +92,5 @@ router.delete('/:comment_id', checkCommentOwnership, function(req, res) {
     }
   });
 });
-
-// middleware to check if the user is logged in
-function isLoggedIn(req, res, next) {
-  if (req.isAuthenticated()) {
-    return next();
-  }
-  res.redirect('/login');
-}
-
-// middleware to check if a user owns the comment
-function checkCommentOwnership(req, res, next) {
-  if (req.isAuthenticated()) {
-    Comment.findById(req.params.comment_id, function(err, foundComment) {
-      if (err) {
-        console.log('Error with edit comment route ', err);
-        res.redirect('back');
-      } else {
-        // does user own the comment
-        if (foundComment.author.id.equals(req.user._id)) {
-          next();
-        } else {
-          res.redirect('back');
-        }
-      }
-    });
-  } else {
-    res.redirect('back');
-  }
-}
 
 module.exports = router;
